Moved the following duplicated functions/classes to a common file hard_sigmoid Round3 binary_tanh_unit binary_sigmoid_unit binarization ternarization DenseLayer Conv2DLayer
