+++ BinaryConnect/svhn.py
@@ -29,7 +29,7 @@ np.random.seed(1234) # for reproducibility?
 # theano.sandbox.cuda.use('gpu1')
 import theano
 import theano.tensor as T
-
+import common
 import lasagne
 
 import cPickle as pickle
@@ -120,7 +120,7 @@ if __name__ == "__main__":
             input_var=input)
 
     # 64C3-64C3-P2
-    cnn = binary_connect.Conv2DLayer(
+    cnn = common.Conv2DLayer(
             cnn,
             binary=binary,
             stochastic=stochastic,
@@ -137,7 +137,7 @@ if __name__ == "__main__":
             alpha=alpha,
             nonlinearity=lasagne.nonlinearities.rectify)
 
-    cnn = binary_connect.Conv2DLayer(
+    cnn = common.Conv2DLayer(
             cnn,
             binary=binary,
             stochastic=stochastic,
@@ -157,7 +157,7 @@ if __name__ == "__main__":
             nonlinearity=lasagne.nonlinearities.rectify)
 
     # 128C3-128C3-P2
-    cnn = binary_connect.Conv2DLayer(
+    cnn = common.Conv2DLayer(
             cnn,
             binary=binary,
             stochastic=stochastic,
@@ -174,7 +174,7 @@ if __name__ == "__main__":
             alpha=alpha,
             nonlinearity=lasagne.nonlinearities.rectify)
 
-    cnn = binary_connect.Conv2DLayer(
+    cnn = common.Conv2DLayer(
             cnn,
             binary=binary,
             stochastic=stochastic,
@@ -194,7 +194,7 @@ if __name__ == "__main__":
             nonlinearity=lasagne.nonlinearities.rectify)
 
     # 256C3-256C3-P2
-    cnn = binary_connect.Conv2DLayer(
+    cnn = common.Conv2DLayer(
             cnn,
             binary=binary,
             stochastic=stochastic,
@@ -211,7 +211,7 @@ if __name__ == "__main__":
             alpha=alpha,
             nonlinearity=lasagne.nonlinearities.rectify)
 
-    cnn = binary_connect.Conv2DLayer(
+    cnn = common.Conv2DLayer(
             cnn,
             binary=binary,
             stochastic=stochastic,
@@ -233,7 +233,7 @@ if __name__ == "__main__":
     # print(cnn.output_shape)
 
     # 1024FP-1024FP-10FP
-    cnn = binary_connect.DenseLayer(
+    cnn = common.DenseLayer(
                 cnn,
                 binary=binary,
                 stochastic=stochastic,
@@ -248,7 +248,7 @@ if __name__ == "__main__":
             alpha=alpha,
             nonlinearity=lasagne.nonlinearities.rectify)
 
-    cnn = binary_connect.DenseLayer(
+    cnn = common.DenseLayer(
                 cnn,
                 binary=binary,
                 stochastic=stochastic,
@@ -263,7 +263,7 @@ if __name__ == "__main__":
             alpha=alpha,
             nonlinearity=lasagne.nonlinearities.rectify)
 
-    cnn = binary_connect.DenseLayer(
+    cnn = common.DenseLayer(
                 cnn,
                 binary=binary,
                 stochastic=stochastic,
